+++ taxel-xml/src/remove_tag_values.rs
